refactor(SingletonBehaviour)!: move to Foundation.Singletons namespace
Move SingletonBehaviour<T> and SingletonRuntime from `Foundation`
to `Foundation.Singletons` namespace for better organization.

This follows .NET naming guidelines and prepares for future
expansion of the Foundation layer with other infrastructure
components (e.g., Pooling, Events).

BREAKING CHANGE: Update using directives from `using Foundation;`
to `using Foundation.Singletons;`
